Refactor Extensions.kt
